# ux/documentation/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/schema-org/search": {
      "get": {
        "tags": [
          "schema-org"
        ],
        "summary": "Semantic Search",
        "operationId": "semantic_search_api_schema_org_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "search_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "both",
              "title": "Search Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "similarity_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.7,
              "title": "Similarity Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/nlp_analysis/reference/dbpedia/resource": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Dbpedia Get Resource",
        "description": "Retrieve structured data from a DBpedia resource URL",
        "operationId": "dbpedia_get_resource_api_nlp_analysis_reference_dbpedia_resource_get",
        "parameters": [
          {
            "name": "resource_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "DBpedia resource URL",
              "title": "Resource Url"
            },
            "description": "DBpedia resource URL"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResponseFormat",
              "description": "Response format",
              "default": "json"
            },
            "description": "Response format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DBpediaResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/dbpedia/search": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Dbpedia Search",
        "description": "Search DBpedia using the search API",
        "operationId": "dbpedia_search_api_nlp_analysis_reference_dbpedia_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query",
              "title": "Query"
            },
            "description": "Search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Result offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Result offset"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResponseFormat",
              "description": "Response format",
              "default": "json"
            },
            "description": "Response format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DBpediaSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/dbpedia/sparql": {
      "post": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Dbpedia Sparql",
        "description": "Execute SPARQL query against DBpedia",
        "operationId": "dbpedia_sparql_api_nlp_analysis_reference_dbpedia_sparql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DBpediaSparqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DBpediaSparqlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/conceptnet/query": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Conceptnet Query",
        "description": "Query ConceptNet with various parameters",
        "operationId": "conceptnet_query_api_nlp_analysis_reference_conceptnet_query_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Starting concept",
              "title": "Start"
            },
            "description": "Starting concept"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ending concept",
              "title": "End"
            },
            "description": "Ending concept"
          },
          {
            "name": "node",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any concept",
              "title": "Node"
            },
            "description": "Any concept"
          },
          {
            "name": "rel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Relation type",
              "title": "Rel"
            },
            "description": "Relation type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Result limit",
              "default": 20,
              "title": "Limit"
            },
            "description": "Result limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Result offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Result offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptNetQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/conceptnet/concept/{concept_path}": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Conceptnet Get Concept",
        "description": "Get data for a specific ConceptNet concept",
        "operationId": "conceptnet_get_concept_api_nlp_analysis_reference_conceptnet_concept__concept_path__get",
        "parameters": [
          {
            "name": "concept_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ConceptNet concept path",
              "title": "Concept Path"
            },
            "description": "ConceptNet concept path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptNetConceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/conceptnet/related/{concept_path}": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Conceptnet Get Related",
        "description": "Get related concepts from ConceptNet",
        "operationId": "conceptnet_get_related_api_nlp_analysis_reference_conceptnet_related__concept_path__get",
        "parameters": [
          {
            "name": "concept_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ConceptNet concept path",
              "title": "Concept Path"
            },
            "description": "ConceptNet concept path"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter for related concepts",
              "title": "Filter"
            },
            "description": "Filter for related concepts"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Result limit",
              "default": 20,
              "title": "Limit"
            },
            "description": "Result limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptNetRelatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/wikidata/sparql": {
      "post": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Wikidata Sparql",
        "description": "Execute SPARQL query against Wikidata",
        "operationId": "wikidata_sparql_api_nlp_analysis_reference_wikidata_sparql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikidataSparqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikidataSparqlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/wikidata/entity": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Wikidata Get Entity",
        "description": "Get structured data for a Wikidata entity",
        "operationId": "wikidata_get_entity_api_nlp_analysis_reference_wikidata_entity_get",
        "parameters": [
          {
            "name": "entity_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Wikidata entity URL",
              "title": "Entity Url"
            },
            "description": "Wikidata entity URL"
          },
          {
            "name": "properties",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated property IDs",
              "title": "Properties"
            },
            "description": "Comma-separated property IDs"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResponseFormat",
              "description": "Response format",
              "default": "json"
            },
            "description": "Response format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikidataEntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/schema-org/entity/{identifier}": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Schema Org Get Entity",
        "description": "Get Schema.org entity with properties and inheritance",
        "operationId": "schema_org_get_entity_api_nlp_analysis_reference_schema_org_entity__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Schema.org entity identifier",
              "title": "Identifier"
            },
            "description": "Schema.org entity identifier"
          },
          {
            "name": "include_inherited",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include inherited properties",
              "default": true,
              "title": "Include Inherited"
            },
            "description": "Include inherited properties"
          },
          {
            "name": "include_children",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include child entities",
              "default": false,
              "title": "Include Children"
            },
            "description": "Include child entities"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaOrgEntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nlp_analysis/reference/schema-org/property/{identifier}": {
      "get": {
        "tags": [
          "nlp-reference",
          "enrichment"
        ],
        "summary": "Schema Org Get Property",
        "description": "Get Schema.org property definition and usage",
        "operationId": "schema_org_get_property_api_nlp_analysis_reference_schema_org_property__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Schema.org property identifier",
              "title": "Identifier"
            },
            "description": "Schema.org property identifier"
          },
          {
            "name": "include_usage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include entities using this property",
              "default": true,
              "title": "Include Usage"
            },
            "description": "Include entities using this property"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaOrgPropertyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DBpediaResourceResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the request was successful"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "description": "Source that provided the data"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At",
            "description": "Timestamp when data was retrieved"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if success is false"
          },
          "resource_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Uri"
          },
          "data_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Url"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success",
          "source",
          "retrieved_at"
        ],
        "title": "DBpediaResourceResponse",
        "description": "Response model for DBpedia resource data"
      },
...
      "DBpediaSparqlRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "SPARQL query"
          },
          "format": {
            "$ref": "#/components/schemas/ResponseFormat",
            "description": "Response format",
            "default": "json"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DBpediaSparqlRequest",
        "description": "Request model for DBpedia SPARQL query"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseFormat": {
        "type": "string",
        "enum": [
          "json",
          "xml",
          "rdf"
        ],
        "title": "ResponseFormat"
      },
...
      "SearchResult": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SchemaOrgEntityOut"
                },
                {
                  "$ref": "#/components/schemas/SchemaOrgPropertyOut"
                }
              ]
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "limit",
          "offset",
          "query"
        ],
        "title": "SearchResult"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "dbpedia",
          "conceptnet",
          "wikidata",
          "schema_org"
        ],
        "title": "SourceType"
      },
...
      "WikidataSparqlRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "SPARQL query"
          },
          "format": {
            "$ref": "#/components/schemas/ResponseFormat",
            "description": "Response format",
            "default": "json"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "WikidataSparqlRequest",
        "description": "Request model for Wikidata SPARQL query"
      },
...
    }
  }
}